requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.2
django-autoslug==1.9.9
django-ckeditor-5==0.2.18
django-cleanup==9.0.0
django-ipware==7.0.1
gunicorn==23.0.0
packaging==25.0
pillow==11.2.1
psycopg==3.2.9
psycopg-binary==3.2.9
python-dotenv==1.1.0
python-ipware==3.0.0
sqlparse==0.5.3
typing_extensions==4.13.2
whitenoise==6.9.0
